+++ packages/eslint-config-vue3-ts/index.js
@@ -14,39 +14,5 @@ module.exports = {
     '@provincial-platform/eslint-config-js',
     '@provincial-platform/eslint-config-vue3',
     '@provincial-platform/eslint-config-ts'
-  ],
-  rules: {
-    '@typescript-eslint/naming-convention': [
-      'error',
-      {
-        selector: 'default',
-        format: ['camelCase'],
-        leadingUnderscore: 'allow',
-        trailingUnderscore: 'allow',
-      },
-      {
-        selector: 'variable',
-        format: ['camelCase', 'UPPER_CASE', 'PascalCase'],
-        leadingUnderscore: 'allow',
-        trailingUnderscore: 'allow',
-      },
-      {
-        selector: 'typeLike',
-        format: ['PascalCase'],
-      },
-      {
-        selector: 'enumMember',
-        format: ['UPPER_CASE'],
-      },
-      // vue components key值可使用PascalCase风格
-      {
-        selector: 'objectLiteralProperty',
-        format: ['camelCase', 'PascalCase', 'snake_case'],
-      },
-      {
-        selector: 'typeProperty',
-        format: ['camelCase', 'PascalCase', 'snake_case', 'UPPER_CASE'],
-      },
-    ],
-  }
+  ]
 }